refactor: move iconmixin to icon
