+++ src/components/alert/alert/alert.ts
@@ -2,20 +2,16 @@ import { spread } from '@open-wc/lit-helpers';
 import { type CSSResultGroup, html, LitElement, nothing, type TemplateResult } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 
-import {
-  LanguageController,
-  type LinkTargetType,
-  SbbIconNameMixin,
-} from '../../core/common-behaviors';
+import { LanguageController, type LinkTargetType } from '../../core/common-behaviors';
 import { EventEmitter } from '../../core/eventing';
 import { i18nCloseAlert, i18nFindOutMore } from '../../core/i18n';
+import { SbbIconNameMixin } from '../../icon';
 import type { SbbTitleLevel } from '../../title';
 
 import style from './alert.scss?lit&inline';
 
 import '../../button/transparent-button';
 import '../../divider';
-import '../../icon';
 import '../../link';
 import '../../title';
