+++ src/components/icon/index.ts
@@ -2,3 +2,4 @@ export * from './icon';
 export * from './icon-base';
 export * from './icon-request';
 export * from './icon-validate';
+export * from './icon-name-mixin';
