+++ src/components/tabs/tab-title/tab-title.ts
@@ -3,14 +3,10 @@ import { LitElement } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 import { html, unsafeStatic } from 'lit/static-html.js';
 
-import {
-  NamedSlotStateController,
-  SbbDisabledMixin,
-  SbbIconNameMixin,
-} from '../../core/common-behaviors';
+import { NamedSlotStateController, SbbDisabledMixin } from '../../core/common-behaviors';
+import { SbbIconNameMixin } from '../../icon';
 import type { SbbTitleLevel } from '../../title';
 
-import '../../icon';
 import style from './tab-title.scss?lit&inline';
 
 /**
